# services_openapi_requirements.yaml
openapi: 3.0.0
info:
  title: Gacha System - OpenAPI 3.0
  description: >-
    API for managing gacha items, auctions, PVP battles, currency, and user
    authentication.


    Useful links:

    - [Project
    repository](https://github.com/LeonardoPantani/UNIPI-GachaAndGames)
  version: 2.0.0
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html#license-text
  contact:
    email: support@gachaandgames.it
  termsOfService: https://localhost/tos
servers:
  - url: https://localhost
    description: Local server
paths:
...
  /auction/internal/is_open_by_item_uuid:
    get:
      tags:
        - auction_internal
      description: Returns true if an open auction is found, false otherwise.
      operationId: is_open_by_item_uuid
      parameters:
        - name: session
          in: cookie
          required: false
          schema:
            type: string 
        - name: uuid
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ItemId'      
      responses:
        '200':
          description: auction.
          content:
            application/json:
              schema:
                type: object
                properties:
                  found: 
                    type: boolean
        '400':
          description: Invalid request.
        '403':
          description: Not authorized.
        '503':
          description: Service temporarily unavailable. Please try again later.
...
  /inventory/internal/get_by_item_uuid:
    get:
      tags:
        - inventory_internal
      description: Returns item with requested uuid.
      operationId: get_item_by_uuid
      parameters:
        - name: session
          in: cookie
          required: false
          schema:
            type: string
        - name: uuid
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ItemId'
      responses:
        '200':
          description: Item info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '400':
          description: Invalid request.
        '403':
          description: Not authorized.
        '404':
          description: Item not found.
        '503':
          description: Service temporarily unavailable. Please try again later.    
...
  /inventory/internal/list_inventory_items:
    get:
      tags:
        - inventory_internal
      description: Returns inventory items owned by user with UUID requested, paginated.
      operationId: get_inventory_items_by_owner_uuid
      parameters:
        - name: session
          in: cookie
          required: false
          schema:
            type: string
        - name: uuid
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserId'
        - name: page_number
          in: query
          description: Page number of the list.
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
      responses:
        '200':
          description: Item list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
        '400':
          description: Invalid request.
        '403':
          description: Not authorized.
        '503':
          description: Service unavailable. Please try again later.
  /inventory/internal/remove:
    delete:
      tags:
        - inventory_internal
      description: Removes item from inventory, by item and owner UUID.
      operationId: remove_item
      parameters:
        - name: session
          in: cookie
          required: false
          schema:
            type: string
        - name: item_uuid
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ItemId'
        - name: owner_uuid
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserId'
      responses:
        '200':
          description: Item deleted.
        '400':
          description: Invalid request.
        '403':
          description: Not authorized.
        '503':
          description: Service unavailable. Please try again later.
...
components:
  schemas:
...
    UserId:
      type: string
      format: uuid
      description: UUID of user.
...
    ItemId:
      type: string
      format: uuid
      description: UUID of single item.
...
    InventoryItem:
      type: object
      properties:
        owner_id:
          $ref: '#/components/schemas/UserId'
        item_id:
          $ref: '#/components/schemas/ItemId'
        gacha_uuid:
          type: string
          format: uuid
          description: UUID representing a gacha
        obtained_date:
          type: string
          format: date-time
          description: Timestamp of latest obtainment
        owners_no:
          type: integer
          description: Number of previous owners of this item
        price_paid:
          type: number
          format: float
          description: Game currency used to buy this item
...